IN-236 refactor: refactoring
