+++ apps/admin/src/application/providers/AuthProvider.tsx
@@ -1,26 +1,9 @@
-import { ReactNode, createContext, useContext, useEffect, useState } from 'react'
+import { ReactNode, useEffect, useState } from 'react'
 import { useCookies } from 'react-cookie'
 
+import { AuthContext } from '@/common/lib/hooks/useAuth'
 import { LoaderBlock } from '@packages/shared/ui/loader/LoaderBlock'
 import { useRouter } from 'next/router'
-
-type AuthContextType = {
-  isAuthenticated: boolean
-  login: (email: string, password: string) => void
-  logout: () => void
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined)
-
-export const useAuth = () => {
-  const context = useContext(AuthContext)
-
-  if (!context) {
-    throw new Error('useAuth must be used within an AuthProvider')
-  }
-
-  return context
-}
 
 type AuthProviderProps = {
   children: ReactNode
