+++ apps/admin/src/common/lib/hooks/useAuth.ts
@@ -1,0 +1,17 @@
+import { createContext, useContext } from 'react'
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined)
+type AuthContextType = {
+  isAuthenticated: boolean
+  login: (email: string, password: string) => void
+  logout: () => void
+}
+export const useAuth = () => {
+  const context = useContext(AuthContext)
+
+  if (!context) {
+    throw new Error('useAuth must be used within an AuthProvider')
+  }
+
+  return context
+}
