+++ apps/admin/src/common/lib/utils/typeGenerator.ts
