+++ apps/admin/src/features/auth/model/useSignIn.ts
@@ -1,8 +1,8 @@
 import { useState } from 'react'
 import { useForm } from 'react-hook-form'
 
-import { useAuth } from '@/application/providers/AuthProvider'
-import { LOGIN_ADMIN } from '@/features/auth/model/loginAdminQueries'
+import { useAuth } from '@/common/lib/hooks/useAuth'
+import { LOGIN_ADMIN } from '@/features/auth/api/loginAdminQueries'
 import { useMutation } from '@apollo/client'
 import { useScopedTranslation } from '@packages/shared/hooks'
