+++ apps/admin/src/features/auth/ui/sign-in-form.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/vars.scss';
+@import '../../../styles/vars';
 
 .card {
   display: flex;
